requirements.txt
google_auth_oauthlib
google-api-python-client
pandas
jupyterlab